Move document code lens to their own modules
